Committed R project on healthcare cost analysis
A nationwide survey of hospital costs conducted by the US Agency for Healthcare consists of 
hospital records of inpatient samples. The given data is restricted to the city of Wisconsin and 
relates to patients in the age group 0-17 years. The agency wants to analyze the data to 
research on the healthcare costs and their utilization